feat: split cli.py into the cli/ command package
