--- cafleet/src/cafleet/cli.py
@@ -1,1315 +1,0 @@
-"""cafleet CLI."""
-
-import contextlib
-import functools
-import importlib.resources
-import json
-import os
-from collections.abc import Callable
-from pathlib import Path
-from typing import NoReturn
-
-import click
-from alembic import command
-from alembic.config import Config
-from alembic.runtime.migration import MigrationContext
-from alembic.script import ScriptDirectory
-from sqlalchemy import create_engine, inspect
-from sqlalchemy.engine.url import make_url
-
-from cafleet import broker, output
-from cafleet.coding_agent import CODING_AGENTS
-from cafleet.config import settings
-from cafleet.multiplexer import MULTIPLEXERS, TmuxError
-
-_MEMBER_PROMPT_TEMPLATE = (
-    "Member of cafleet fleet {fleet_id} "
-    "(agent={agent_id}, director={director_agent_id}).\n"
-    "Load skill 'cafleet'. Bash auto-approves. Poll: "
-    "cafleet --fleet-id {fleet_id} message poll --agent-id {agent_id}"
-)
-
-
-def _ensure_tmux_or_die() -> None:
-    try:
-        MULTIPLEXERS["tmux"].ensure_available()
-    except TmuxError as exc:
-        raise click.ClickException(str(exc)) from exc
-
-
-_full_flag = click.option("--full", "full", is_flag=True, default=False, hidden=True)
-_quiet_flag = click.option("--quiet", "quiet", is_flag=True, default=False, hidden=True)
-
-
-def _full_flag_with_help(help_text: str):
-    return click.option(
-        "--full", "full", is_flag=True, default=False, hidden=True, help=help_text
-    )
-
-
-def _quiet_flag_with_help(help_text: str):
-    return click.option(
-        "--quiet", "quiet", is_flag=True, default=False, hidden=True, help=help_text
-    )
-
-
-def _director_member_options(func):
-    return click.option(
-        "--member-id", type=int, required=True, help="Target member's agent ID"
-    )(func)
-
-
-def _require_fleet_id(ctx: click.Context) -> None:
-    if ctx.obj["fleet_id"] is None:
-        raise click.ClickException(
-            "--fleet-id <int> is required for this subcommand. "
-            "Create a fleet with 'cafleet fleet create' and pass its id."
-        )
-
-
-def _client_command(
-    *,
-    requires_agent_fleet: bool = False,
-    text_formatter: Callable[..., str] | None = None,
-    truncates_task_text: bool = False,
-    renders_agent_card: bool = False,
-):
-    """Subsume the boilerplate blocks shared by client subcommands.
-
-    Branches:
-    - ``truncates_task_text=True`` → JSON output goes through
-      ``render_tasks_in_result`` + ``truncate_task_text``; text formatter is
-      called as ``text_formatter(result, full=, quiet=)``.
-    - ``renders_agent_card=True`` → JSON output goes through
-      ``render_agents_in_result``; text formatter is called as
-      ``text_formatter(result, full=)``.
-    - Neither → JSON output is the raw broker result; text formatter is
-      called as ``text_formatter(result)``.
-
-    The two ``renders_*`` flags are mutually exclusive — a command renders
-    tasks OR agent cards, never both.
-    """
-
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(ctx, *args, **kwargs):
-            _require_fleet_id(ctx)
-            fleet_id = ctx.obj["fleet_id"]
-            try:
-                if requires_agent_fleet:
-                    agent_id = kwargs.get("agent_id")
-                    if agent_id is None:
-                        raise click.ClickException(
-                            "_client_command(requires_agent_fleet=True) but no "
-                            "'agent_id' kwarg was passed. Check the @click.option "
-                            "declaration on this command."
-                        )
-                    if not broker.verify_agent_fleet(agent_id, fleet_id):
-                        raise click.ClickException(
-                            f"agent {agent_id} is not a member of fleet {fleet_id}."
-                        )
-                result = func(ctx, *args, **kwargs)
-                full = kwargs.get("full", False)
-                if truncates_task_text:
-                    output.truncate_task_text(result, full=full)
-                    rendered = output.render_tasks_in_result(result, full=full)
-                elif renders_agent_card:
-                    rendered = output.render_agents_in_result(result, full=full)
-                else:
-                    rendered = result
-                if ctx.obj["json_output"]:
-                    click.echo(output.format_json(rendered))
-                elif text_formatter is not None:
-                    if truncates_task_text:
-                        extra = {"quiet": kwargs["quiet"]} if "quiet" in kwargs else {}
-                        click.echo(text_formatter(result, full=full, **extra))
-                    elif renders_agent_card:
-                        click.echo(text_formatter(result, full=full))
-                    else:
-                        click.echo(text_formatter(result))
-                else:
-                    click.echo(output.format_json(rendered))
-            except click.ClickException:
-                raise
-            except Exception as exc:
-                raise click.ClickException(str(exc)) from exc
-            return result
-
-        return wrapper
-
-    return decorator
-
-
-def _sync_db_url() -> str:
-    return str(make_url(settings.database_url).set(drivername="sqlite"))
-
-
-@click.group()
-@click.version_option(package_name="cafleet", message="cafleet %(version)s")
-@click.option(
-    "--json", "json_output", is_flag=True, default=False, help="Output in JSON format"
-)
-@click.option(
-    "--fleet-id",
-    "fleet_id",
-    type=int,
-    default=None,
-    help="Fleet ID (integer); required for client subcommands.",
-)
-@click.pass_context
-def cli(ctx, json_output, fleet_id):
-    """CAFleet — CLI for the message broker and agent registry."""
-    ctx.ensure_object(dict)
-    ctx.obj["fleet_id"] = fleet_id
-    ctx.obj["json_output"] = json_output
-
-
-@cli.group()
-def db() -> None:
-    """Database schema management commands."""
-
-
-@db.command("init")
-def init() -> None:
-    """Initialize or migrate the registry database to the head revision."""
-    sync_url = _sync_db_url()
-    db_file_str = make_url(sync_url).database
-    if not db_file_str:
-        raise click.ClickException("database URL has no file path")
-    db_file = Path(db_file_str)
-
-    db_file.parent.mkdir(parents=True, exist_ok=True)
-
-    # ``as_file`` materializes the bundled ``alembic.ini`` to a real path
-    # because when cafleet is installed from a zipped wheel ``files(...)``
-    # returns a Traversable that Alembic cannot open. Hold the context
-    # open across ``command.upgrade`` so the extracted file survives.
-    with importlib.resources.as_file(
-        importlib.resources.files("cafleet.db") / "alembic.ini"
-    ) as ini_path:
-        cfg = Config(str(ini_path))
-        cfg.set_main_option("sqlalchemy.url", sync_url)
-
-        engine = create_engine(sync_url)
-        try:
-            with engine.connect() as conn:
-                inspector = inspect(conn)
-                tables = set(inspector.get_table_names())
-                has_alembic_version = "alembic_version" in tables
-                non_alembic_tables = tables - {"alembic_version"}
-
-                current_rev: str | None = None
-                if has_alembic_version:
-                    ctx = MigrationContext.configure(conn)
-                    current_rev = ctx.get_current_revision()
-
-            if non_alembic_tables and not has_alembic_version:
-                raise click.ClickException(
-                    "DB has existing tables but no alembic_version. "
-                    "Run `alembic stamp head` manually if you are sure the "
-                    "schema matches."
-                )
-
-            script = ScriptDirectory.from_config(cfg)
-            head_rev = script.get_current_head()
-
-            if current_rev is not None:
-                known_revisions = {rev.revision for rev in script.walk_revisions()}
-                if current_rev not in known_revisions:
-                    raise click.ClickException(
-                        f"DB schema is at revision {current_rev} which "
-                        f"is unknown to this version of cafleet. "
-                        f"Refusing to downgrade automatically."
-                    )
-
-            if current_rev == head_rev:
-                click.echo(f"Already at head ({head_rev}); nothing to do.")
-                return
-
-            old_rev = current_rev or "(empty)"
-            command.upgrade(cfg, "head")
-            if current_rev is None:
-                click.echo(
-                    f"Created {db_file} and applied migrations to head ({head_rev})."
-                )
-            else:
-                click.echo(f"Upgraded from {old_rev} to {head_rev}.")
-        finally:
-            engine.dispose()
-
-
-@cli.group()
-def fleet() -> None:
-    """Fleet management commands."""
-
-
-@fleet.command("create")
-@click.option("--label", default=None, help="Optional human-readable label.")
-@click.option(
-    "--coding-agent",
-    "coding_agent",
-    type=click.Choice(list(CODING_AGENTS.keys())),
-    default="claude",
-    show_default=True,
-    help="Coding-agent binary to spawn / declare for the placement.",
-)
-@click.option("--json", "as_json", is_flag=True, help="Output as JSON.")
-@_full_flag
-@click.pass_context
-def fleet_create(
-    ctx: click.Context,
-    label: str | None,
-    coding_agent: str,
-    as_json: bool,
-    full: bool,
-) -> None:
-    """Create a new fleet (must be run inside a tmux session)."""
-    try:
-        MULTIPLEXERS["tmux"].ensure_available()
-        director_ctx = MULTIPLEXERS["tmux"].context_discovery()
-    except TmuxError as exc:
-        raise click.ClickException(
-            "cafleet fleet create must be run inside a tmux session"
-        ) from exc
-
-    result = broker.create_fleet(
-        label=label,
-        director_context=director_ctx,
-        coding_agent=coding_agent,
-    )
-
-    if as_json or ctx.obj["json_output"]:
-        click.echo(output.format_json(result))
-    else:
-        click.echo(output.format_fleet_create(result, full=full))
-
-
-@fleet.command("list")
-@click.option("--json", "as_json", is_flag=True, help="Output as JSON.")
-@click.pass_context
-def fleet_list(ctx: click.Context, as_json: bool) -> None:
-    """List all fleets."""
-    rows = broker.list_fleets()
-
-    if as_json or ctx.obj["json_output"]:
-        click.echo(output.format_json(rows))
-    else:
-        if not rows:
-            click.echo("No fleets found.")
-            return
-        click.echo(
-            f"{'FLEET_ID':<40} {'DIRECTOR':<40} {'LABEL':<20} "
-            f"{'AGENTS':<8} {'CREATED_AT'}"
-        )
-        for r in rows:
-            click.echo(
-                f"{r['fleet_id']:<40} {r['director_agent_id'] or '':<40} "
-                f"{r['label'] or '':<20} {r['agent_count']:<8} {r['created_at']}"
-            )
-
-
-@fleet.command("show")
-@click.argument("fleet_id", type=int)
-@click.option("--json", "as_json", is_flag=True, help="Output as JSON.")
-@click.pass_context
-def fleet_show(ctx: click.Context, fleet_id: int, as_json: bool) -> None:
-    """Show details of a single fleet."""
-    result = broker.get_fleet(fleet_id)
-    if result is None:
-        raise click.ClickException(f"fleet '{fleet_id}' not found.")
-
-    if as_json or ctx.obj["json_output"]:
-        click.echo(output.format_json(result))
-    else:
-        lines = [
-            f"fleet_id: {result['fleet_id']}",
-            f"label:      {result['label'] or ''}",
-            f"created_at: {result['created_at']}",
-        ]
-        if result["deleted_at"] is not None:
-            lines.append(f"deleted_at: {result['deleted_at']}")
-        click.echo("\n".join(lines))
-
-
-@fleet.command("delete")
-@click.argument("fleet_id", type=int)
-def fleet_delete(fleet_id: int) -> None:
-    """Soft-delete a fleet and deregister every active agent (idempotent)."""
-    result = broker.delete_fleet(fleet_id)
-    n = result["deregistered_count"]
-    click.echo(f"Deleted fleet {fleet_id}. Deregistered {n} agents.")
-
-
-@cli.command("server")
-@click.option(
-    "--host",
-    default=settings.broker_host,
-    show_default=True,
-    help="Bind address.",
-)
-@click.option(
-    "--port",
-    default=settings.broker_port,
-    show_default=True,
-    type=int,
-    help="Bind port.",
-)
-def server(host: str, port: int) -> None:
-    """Start the admin WebUI FastAPI server."""
-    import uvicorn
-
-    uvicorn.run(
-        "cafleet.webui.app:app",
-        host=host,
-        port=port,
-    )
-
-
-@cli.command("doctor")
-@click.pass_context
-def doctor(ctx) -> None:
-    """Print the calling pane's tmux session/window/pane identifiers."""
-    _ensure_tmux_or_die()
-
-    try:
-        director_ctx = MULTIPLEXERS["tmux"].context_discovery()
-    except TmuxError as exc:
-        raise click.ClickException(str(exc)) from exc
-
-    tmux_pane_env = os.environ["TMUX_PANE"]
-
-    if ctx.obj["json_output"]:
-        click.echo(
-            output.format_json(
-                {
-                    "tmux": {
-                        "session_name": director_ctx.session,
-                        "window_id": director_ctx.window_id,
-                        "pane_id": director_ctx.pane_id,
-                        "tmux_pane_env": tmux_pane_env,
-                    }
-                },
-            )
-        )
-    else:
-        click.echo("tmux:")
-        click.echo(f"  session_name:  {director_ctx.session}")
-        click.echo(f"  window_id:     {director_ctx.window_id}")
-        click.echo(f"  pane_id:       {director_ctx.pane_id}")
-        click.echo(f"  TMUX_PANE:     {tmux_pane_env}")
-
-
-@cli.group()
-def agent() -> None:
-    """Agent registry commands."""
-
-
-@cli.group()
-def message() -> None:
-    """Message broker commands."""
-
-
-@agent.command("register")
-@click.option("--name", required=True, help="Agent name")
-@click.option("--description", required=True, help="Agent description")
-@click.option("--skills", default=None, help="Skills as JSON string")
-@click.pass_context
-@_client_command(text_formatter=output.format_register)
-def agent_register(ctx, name, description, skills):
-    """Register a new agent with the broker."""
-    parsed_skills = None
-    if skills is not None:
-        try:
-            parsed_skills = json.loads(skills)
-        except json.JSONDecodeError as exc:
-            raise click.ClickException(f"Invalid JSON in --skills: {exc}") from exc
-    return broker.register_agent(
-        ctx.obj["fleet_id"],
-        name,
-        description,
-        skills=parsed_skills,
-    )
-
-
-@message.command("send")
-@click.option("--agent-id", type=int, required=True, help="Agent ID")
-@click.option("--to", type=int, required=True, help="Recipient agent ID")
-@click.option("--text", required=True, help="Message text")
-@_full_flag
-@_quiet_flag
-@click.pass_context
-@_client_command(
-    requires_agent_fleet=True,
-    text_formatter=lambda r, *, full, quiet: (
-        str(r["task"]["task_id"])
-        if quiet
-        else "Message sent.\n" + output.format_task(r, full=full)
-    ),
-    truncates_task_text=True,
-)
-def message_send(ctx, agent_id, to, text, full, quiet):
-    """Send a unicast message to another agent."""
-    fleet_id = ctx.obj["fleet_id"]
-    return broker.send_message(
-        fleet_id,
-        agent_id,
-        to,
-        text,
-    )
-
-
-@message.command("broadcast")
-@click.option("--agent-id", type=int, required=True, help="Agent ID")
-@click.option("--text", required=True, help="Message text")
-@_full_flag
-@click.pass_context
-@_client_command(
-    text_formatter=lambda r, *, full: (
-        output.format_task(r[0]["task"], full=True)
-        if full
-        else f"broadcast id={r[0]['task']['task_id']} "
-        f"recipients={r[0]['notifications_sent_count']}"
-    ),
-    truncates_task_text=True,
-)
-def message_broadcast(ctx, agent_id, text, full):
-    """Broadcast a message to all agents."""
-    return broker.broadcast_message(
-        ctx.obj["fleet_id"],
-        agent_id,
-        text,
-    )
-
-
-@message.command("poll")
-@click.option("--agent-id", type=int, required=True, help="Agent ID")
-@_full_flag_with_help("Disable body truncation.")
-@click.pass_context
-@_client_command(
-    requires_agent_fleet=True,
-    text_formatter=lambda r, *, full: output.format_indexed_list(
-        r, lambda t: output.format_task(t, full=full), "No messages found."
-    ),
-    truncates_task_text=True,
-)
-def message_poll(ctx, agent_id, full):
-    """Poll inbox for un-acked messages."""
-    return broker.poll_tasks(agent_id)
-
-
-@message.command("ack")
-@click.option("--agent-id", type=int, required=True, help="Agent ID")
-@click.option("--task-id", type=int, required=True, help="Task ID to acknowledge")
-@_full_flag_with_help("Disable body truncation.")
-@_quiet_flag_with_help("Print only the task id.")
-@click.pass_context
-@_client_command(
-    requires_agent_fleet=True,
-    text_formatter=lambda r, *, full, quiet: (
-        str(r["task"]["task_id"])
-        if quiet
-        else "Message acknowledged.\n" + output.format_task(r, full=full)
-    ),
-    truncates_task_text=True,
-)
-def message_ack(ctx, agent_id, task_id, full, quiet):
-    """Acknowledge receipt of a message."""
-    return broker.ack_task(agent_id, task_id)
-
-
-@message.command("cancel")
-@click.option("--agent-id", type=int, required=True, help="Agent ID")
-@click.option("--task-id", type=int, required=True, help="Task ID to cancel")
-@_full_flag_with_help("Disable body truncation.")
-@click.pass_context
-@_client_command(
-    requires_agent_fleet=True,
-    text_formatter=lambda r, *, full: (
-        "Task canceled.\n" + output.format_task(r, full=full)
-    ),
-    truncates_task_text=True,
-)
-def message_cancel(ctx, agent_id, task_id, full):
-    """Cancel (retract) a sent message."""
-    return broker.cancel_task(agent_id, task_id)
-
-
-@message.command("show")
-@click.option("--agent-id", type=int, required=True, help="Agent ID")
-@click.option("--task-id", type=int, required=True, help="Task ID to retrieve")
-@_full_flag_with_help("Disable body truncation.")
-@click.pass_context
-@_client_command(
-    requires_agent_fleet=True,
-    text_formatter=lambda r, *, full: output.format_task(r, full=full),
-    truncates_task_text=True,
-)
-def message_show(ctx, agent_id, task_id, full):
-    """Get details of a specific task."""
-    fleet_id = ctx.obj["fleet_id"]
-    return broker.get_task(fleet_id, task_id)
-
-
-@agent.command("list")
-@_full_flag
-@click.pass_context
-@_client_command(
-    text_formatter=lambda agents, *, full: output.format_indexed_list(
-        agents, lambda a: output.format_agent(a, full=full), "No agents found."
-    ),
-    renders_agent_card=True,
-)
-def agent_list(ctx, full):
-    """List registered agents in the fleet."""
-    return broker.list_agents(ctx.obj["fleet_id"])
-
-
-@agent.command("show")
-@click.option("--agent-id", type=int, required=True, help="Agent ID")
-@click.option(
-    "--id", "target_agent_id", type=int, required=True, help="Target agent ID"
-)
-@_full_flag
-@click.pass_context
-@_client_command(
-    requires_agent_fleet=True,
-    text_formatter=output.format_agent,
-    renders_agent_card=True,
-)
-def agent_show(ctx, agent_id, target_agent_id, full):
-    """Show detail for a specific agent."""
-    fleet_id = ctx.obj["fleet_id"]
-    result = broker.get_agent(target_agent_id, fleet_id)
-    if result is None:
-        raise click.ClickException(f"Agent {target_agent_id} not found")
-    return result
-
-
-@agent.command("deregister")
-@click.option("--agent-id", type=int, required=True, help="Agent ID")
-@click.pass_context
-@_client_command(
-    requires_agent_fleet=True,
-    text_formatter=lambda _: "Agent deregistered successfully.",
-)
-def agent_deregister(ctx, agent_id):
-    """Deregister this agent from the broker."""
-    deregistered = broker.deregister_agent(agent_id)
-    if not deregistered:
-        raise click.ClickException(
-            f"agent {agent_id} not found or already deregistered."
-        )
-    return {"status": "deregistered"}
-
-
-@cli.group()
-def member():
-    """Manage tmux-backed member agents (Director only)."""
-
-
-_PLACEMENT_MISSING_DEFAULT = (
-    "agent {member_id} has no placement row; it was not "
-    "spawned via `cafleet member create`."
-)
-
-
-def _require_member_pane(placement: dict, member_id: int, action: str) -> str:
-    pane_id = placement["tmux_pane_id"]
-    if pane_id is None:
-        raise click.ClickException(
-            f"member {member_id} has no pane yet (pending placement) "
-            f"— nothing to {action}."
-        )
-    return pane_id
-
-
-def _load_authorized_member(
-    fleet_id: int,
-    member_id: int,
-    *,
-    placement_missing_template: str = _PLACEMENT_MISSING_DEFAULT,
-) -> tuple[dict, dict]:
-    """Resolve a fleet-scoped member's agent + placement.
-
-    The only boundary is fleet isolation: ``broker.get_agent(member_id,
-    fleet_id)`` returns ``None`` for a ``member_id`` that is not an active agent
-    of ``fleet_id``, so a cross-fleet / unknown / inactive target raises "not
-    found". There is no caller-auth check — any active in-fleet agent with a
-    placement row (the root Director included) is a valid target; an in-fleet
-    agent without a placement row raises the placement-missing error below.
-
-    ``placement_missing_template`` is a ``{member_id}`` format string for the
-    "no placement" path, because each caller points users at a different
-    follow-up command (``cafleet member create`` by default; ``cafleet agent
-    deregister`` from delete). Pane-id presence is NOT checked here — delete
-    tolerates a pending placement while the others reject it.
-
-    Callers MUST use ``target["agent_id"]``, since reassigning the
-    ``member_id`` local param does not propagate to the caller.
-    """
-    try:
-        target = broker.get_agent(member_id, fleet_id)
-    except Exception as exc:
-        raise click.ClickException(f"failed to fetch member: {exc}") from exc
-    if target is None:
-        raise click.ClickException(f"Agent {member_id} not found")
-    placement = target["placement"]
-    if placement is None:
-        raise click.ClickException(
-            placement_missing_template.format(member_id=member_id)
-        )
-    return target, placement
-
-
-def _read_prompt_file(path: str) -> str:
-    """Read the spawn prompt from a file, validating absolute path / readability / UTF-8 / non-empty.
-
-    Owns the five error surfaces: relative path →
-    ``UsageError``; missing / non-regular file → ``ClickException``;
-    permission / generic-I/O failures → ``ClickException``; invalid UTF-8 →
-    ``ClickException``; zero-byte or whitespace-only contents →
-    ``ClickException``. Returns the file body verbatim — no stripping,
-    trailing newlines preserved.
-
-    The existence and regular-file checks ride entirely on the
-    ``read_bytes()`` exception surface (no ``is_file()`` pre-check), so a
-    path on which we lack traverse permission cannot leak through a stat
-    gate and surface as the wrong message — ``PermissionError`` from
-    ``read_bytes`` lands in the ``file is not readable`` branch directly.
-    """
-    if not Path(path).is_absolute():
-        raise click.UsageError(
-            f"--prompt-file requires an absolute path (got '{path}'). "
-            "Resolve relative paths against your BASE first — "
-            "see the `cafleet-base-dir` skill."
-        )
-    file_path = Path(path)
-    try:
-        # read_bytes() + decode() instead of read_text() so universal-newline
-        # translation does NOT collapse CRLF / CR to LF — the success-criterion
-        # promises the file body lands in the spawn argv byte-for-byte verbatim.
-        content = file_path.read_bytes().decode("utf-8")
-    except (FileNotFoundError, IsADirectoryError) as exc:
-        # FileNotFoundError → the path does not name an existing file.
-        # IsADirectoryError → the path names a directory, not a regular file.
-        # Both map to the § 6 "missing or non-regular file" surface.
-        raise click.ClickException(
-            f"--prompt-file {path}: file does not exist or is not a regular file."
-        ) from exc
-    except PermissionError as exc:
-        raise click.ClickException(
-            f"--prompt-file {path}: file is not readable."
-        ) from exc
-    except UnicodeDecodeError as exc:
-        raise click.ClickException(
-            f"--prompt-file {path}: file is not valid UTF-8."
-        ) from exc
-    except OSError as exc:
-        # Reserved for true I/O failures (EIO, ENOSPC, EBUSY, …). The earlier
-        # FileNotFoundError / IsADirectoryError / PermissionError branches
-        # handle the OSError subclasses that have a more precise § 6 message.
-        raise click.ClickException(
-            f"--prompt-file {path}: file is not readable."
-        ) from exc
-    if content == "" or content.isspace():
-        raise click.ClickException(f"--prompt-file {path}: file is empty.")
-    return content
-
-
-def _resolve_prompt(
-    ctx: click.Context,
-    director_agent_id: int,
-    new_agent_id: int,
-    prompt_argv: tuple[str, ...],
-    prompt_file: str | None = None,
-) -> str:
-    """Substitute ``fleet_id`` / ``agent_id`` / ``director_agent_id`` into the spawn prompt.
-
-    Runs ``str.format`` on the chosen template (file > positional > default)
-    so custom prompts must double literal braces (``{{`` / ``}}``) to survive
-    the substitution.
-    """
-    fleet_id = ctx.obj["fleet_id"]
-    if prompt_file is not None:
-        template = _read_prompt_file(prompt_file)
-    elif prompt_argv:
-        template = " ".join(prompt_argv)
-    else:
-        template = _MEMBER_PROMPT_TEMPLATE
-    try:
-        return template.format(
-            fleet_id=fleet_id,
-            agent_id=new_agent_id,
-            director_agent_id=director_agent_id,
-        )
-    except KeyError as exc:
-        raise click.UsageError(
-            f"Unknown placeholder {exc} in custom prompt. "
-            "Supported placeholders: {fleet_id}, {agent_id}, "
-            "{director_agent_id}. "
-            "Double literal braces ({{, }}) to keep them as text."
-        ) from exc
-    except (ValueError, IndexError, AttributeError) as exc:
-        raise click.UsageError(
-            f"Malformed custom prompt: {exc}. "
-            "Double literal braces ({{, }}) to keep them as text."
-        ) from exc
-
-
-def _deregister_with_warning(new_agent_id: int, *, fleet_id: int) -> None:
-    """Best-effort deregister; emit warning to stderr if it fails."""
-    try:
-        broker.deregister_agent(new_agent_id)
-    except Exception as drop_exc:
-        click.echo(
-            f"WARNING: rollback deregister failed — agent {new_agent_id} is "
-            f"orphaned in the registry. Run `cafleet --fleet-id {fleet_id} "
-            f"agent deregister --agent-id {new_agent_id}` manually to clean up. "
-            f"Cause: {drop_exc}",
-            err=True,
-        )
-
-
-def _rollback_register(new_agent_id: int, *, fleet_id: int, reason: str) -> NoReturn:
-    """Best-effort deregister of a just-created agent, then raise ClickException."""
-    _deregister_with_warning(new_agent_id, fleet_id=fleet_id)
-    raise click.ClickException(f"{reason}. Rolled back registration of {new_agent_id}.")
-
-
-@member.command("create")
-@click.option("--agent-id", type=int, required=True, help="Director's agent ID")
-@click.option("--name", required=True, help="Member name")
-@click.option("--description", required=True, help="Member description")
-@click.option(
-    "--coding-agent",
-    "coding_agent",
-    type=click.Choice(list(CODING_AGENTS.keys())),
-    default="claude",
-    show_default=True,
-    help="Coding-agent binary to spawn / declare for the placement.",
-)
-@click.option(
-    "--model",
-    "model",
-    type=str,
-    default=None,
-    help="Model passed to the backend binary.",
-)
-@click.option(
-    "--prompt-file",
-    "prompt_file",
-    type=str,
-    default=None,
-    help="Read spawn prompt from FILE (abs path, UTF-8).",
-)
-@_full_flag
-@click.argument("prompt_argv", nargs=-1)
-@click.pass_context
-def member_create(
-    ctx,
-    agent_id,
-    name,
-    description,
-    coding_agent,
-    model,
-    prompt_file,
-    full,
-    prompt_argv,
-):
-    """Register a new member and spawn its pane."""
-    if prompt_file is not None and prompt_argv:
-        raise click.UsageError(
-            "--prompt-file and the positional prompt argument are mutually exclusive."
-        )
-    _require_fleet_id(ctx)
-    fleet_id = ctx.obj["fleet_id"]
-
-    agent = CODING_AGENTS[coding_agent]
-
-    try:
-        agent.validate_model(model)
-    except ValueError as exc:
-        raise click.UsageError(str(exc)) from exc
-
-    try:
-        MULTIPLEXERS["tmux"].ensure_available()
-        agent.ensure_available()
-        director_ctx = MULTIPLEXERS["tmux"].context_discovery()
-    except (TmuxError, RuntimeError) as exc:
-        raise click.ClickException(str(exc)) from exc
-
-    try:
-        result = broker.register_agent(
-            fleet_id,
-            name,
-            description,
-            placement={
-                "director_agent_id": agent_id,
-                "tmux_session": director_ctx.session,
-                "tmux_window_id": director_ctx.window_id,
-                "tmux_pane_id": None,
-                "coding_agent": coding_agent,
-            },
-        )
-    except Exception as exc:
-        raise click.ClickException(f"register failed: {exc}") from exc
-    new_agent_id = result["agent_id"]
-
-    try:
-        prompt = _resolve_prompt(ctx, agent_id, new_agent_id, prompt_argv, prompt_file)
-    except (click.UsageError, click.ClickException):
-        # Re-raise unwrapped so the exact message from docs/spec/cli-options.md
-        # § Error Messages reaches the operator. Wrapping via _rollback_register
-        # would prepend "prompt resolution failed:" and append "Rolled back
-        # registration of <id>." (with a stray ".." when the inner message
-        # already ends in a period), and would also downgrade UsageError exit
-        # code 2 → ClickException exit 1.
-        _deregister_with_warning(new_agent_id, fleet_id=fleet_id)
-        raise
-
-    spawn_command = agent.build_spawn_argv(prompt, display_name=name, model=model)
-
-    try:
-        db_url = os.environ.get("CAFLEET_DATABASE_URL")
-        fwd_env = {"CAFLEET_DATABASE_URL": db_url} if db_url else {}
-        pane_id = MULTIPLEXERS["tmux"].split_window(
-            target_window_id=director_ctx.window_id,
-            env=fwd_env,
-            command=spawn_command,
-        )
-    except TmuxError as exc:
-        _rollback_register(
-            new_agent_id,
-            fleet_id=fleet_id,
-            reason=f"tmux split-window failed: {exc}",
-        )
-
-    try:
-        placement_view = broker.update_placement_pane_id(new_agent_id, pane_id)
-    except Exception as exc:
-        # Pane is alive but the registration row is dangling; /exit the pane
-        # and roll back the agent so the caller can retry cleanly.
-        with contextlib.suppress(TmuxError):
-            MULTIPLEXERS["tmux"].send_exit(target_pane_id=pane_id, ignore_missing=True)
-        _rollback_register(
-            new_agent_id,
-            fleet_id=fleet_id,
-            reason=f"placement update failed: {exc}",
-        )
-    if placement_view is None:
-        with contextlib.suppress(TmuxError):
-            MULTIPLEXERS["tmux"].send_exit(target_pane_id=pane_id, ignore_missing=True)
-        _rollback_register(
-            new_agent_id,
-            fleet_id=fleet_id,
-            reason="placement row vanished before pane-id patch",
-        )
-
-    result["placement"] = placement_view
-    if ctx.obj["json_output"]:
-        click.echo(output.format_json(result))
-    else:
-        click.echo(output.format_member(result, full=full))
-
-
-@member.command("delete")
-@_director_member_options
-@click.option(
-    "--force",
-    "-f",
-    "force",
-    is_flag=True,
-    default=False,
-    help="Skip /exit; kill-pane immediately.",
-)
-@click.pass_context
-def member_delete(ctx, member_id, force):
-    """Deregister a member agent and close its tmux pane."""
-    _require_fleet_id(ctx)
-    fleet_id = ctx.obj["fleet_id"]
-
-    _ensure_tmux_or_die()
-
-    fleet = broker.get_fleet(fleet_id)
-    if fleet is not None and member_id == fleet["director_agent_id"]:
-        raise click.ClickException(
-            "cannot deregister the root Director; use 'cafleet fleet delete' instead"
-        )
-
-    target, placement = _load_authorized_member(
-        fleet_id,
-        member_id,
-        placement_missing_template=(
-            "agent {member_id} has no placement; use `cafleet agent deregister` instead"
-        ),
-    )
-    member_id = target["agent_id"]
-    pane_id = placement["tmux_pane_id"]
-
-    if pane_id is None:
-        try:
-            broker.deregister_agent(member_id)
-        except Exception as exc:
-            raise click.ClickException(f"deregister failed: {exc}") from exc
-        pane_status = "(pending — no pane)"
-        _emit_member_delete_output(
-            ctx, member_id, pane_status, header="Member deleted."
-        )
-        return
-
-    if force:
-        try:
-            MULTIPLEXERS["tmux"].kill_pane(target_pane_id=pane_id, ignore_missing=True)
-        except TmuxError as exc:
-            raise click.ClickException(
-                f"kill_pane failed for pane {pane_id}: {exc}. "
-                f"The tmux server may be unreachable. Verify with 'cafleet doctor', "
-                f"then re-run the command."
-            ) from exc
-        try:
-            broker.deregister_agent(member_id)
-        except Exception as exc:
-            raise click.ClickException(f"deregister failed: {exc}") from exc
-        pane_status = f"{pane_id} (killed)"
-        _emit_member_delete_output(
-            ctx, member_id, pane_status, header="Member deleted (--force)."
-        )
-        return
-
-    try:
-        MULTIPLEXERS["tmux"].send_exit(target_pane_id=pane_id, ignore_missing=True)
-    except TmuxError as exc:
-        raise click.ClickException(
-            f"send_exit failed for pane {pane_id}: {exc}. "
-            f"The tmux server may be unreachable. Verify with 'cafleet doctor', "
-            f"then re-run 'cafleet member delete', or use '--force' to kill the "
-            f"pane directly."
-        ) from exc
-
-    try:
-        gone = MULTIPLEXERS["tmux"].wait_for_pane_gone(
-            target_pane_id=pane_id, timeout=15.0, interval=0.5
-        )
-    except TmuxError as exc:
-        raise click.ClickException(
-            f"tmux call failed while waiting for pane {pane_id} to close: {exc}"
-        ) from exc
-
-    if gone:
-        try:
-            broker.deregister_agent(member_id)
-        except Exception as exc:
-            raise click.ClickException(f"deregister failed: {exc}") from exc
-        pane_status = f"{pane_id} (closed)"
-        _emit_member_delete_output(
-            ctx, member_id, pane_status, header="Member deleted."
-        )
-        return
-
-    try:
-        tail = MULTIPLEXERS["tmux"].capture_pane(target_pane_id=pane_id, lines=80)
-    except TmuxError as exc:
-        click.echo(
-            f"Warning: capture_pane failed during timeout handling: {exc}. "
-            f"The timeout error and recovery hint still print.",
-            err=True,
-        )
-        tail = ""
-
-    click.echo(
-        f"Error: pane {pane_id} did not close within 15.0s after /exit.", err=True
-    )
-    click.echo(f"--- pane {pane_id} tail (last 80 lines) ---", err=True)
-    click.echo(tail, err=True)
-    click.echo("---", err=True)
-    click.echo(
-        "Recovery: inspect with `cafleet member capture`, answer any prompt with "
-        "`cafleet member send-input`, then re-run `cafleet member delete`. "
-        "Or re-run with `--force` to skip the wait and kill the pane.",
-        err=True,
-    )
-
-    pane_status = f"{pane_id} (timeout)"
-    if ctx.obj["json_output"]:
-        click.echo(
-            output.format_json(
-                {"agent_id": member_id, "pane_status": pane_status},
-            )
-        )
-    ctx.exit(2)
-
-
-def _emit_member_delete_output(
-    ctx: click.Context,
-    member_id: int,
-    pane_status: str,
-    *,
-    header: str,
-) -> None:
-    if ctx.obj["json_output"]:
-        click.echo(
-            output.format_json(
-                {"agent_id": member_id, "pane_status": pane_status},
-            )
-        )
-    else:
-        click.echo(header)
-        click.echo(f"  agent_id:  {member_id}")
-        click.echo(f"  pane_id:   {pane_status}")
-
-
-@member.command("list")
-@click.option(
-    "--activity",
-    "activity",
-    is_flag=True,
-    default=False,
-    hidden=True,
-)
-@click.pass_context
-def member_list(ctx, activity):
-    """List every member of the fleet (the root Director is excluded)."""
-    _require_fleet_id(ctx)
-    fleet_id = ctx.obj["fleet_id"]
-    try:
-        if activity:
-            rows = broker.list_members_with_activity(fleet_id)
-        else:
-            rows = broker.list_members(fleet_id)
-    except click.ClickException:
-        raise
-    except Exception as exc:
-        raise click.ClickException(str(exc)) from exc
-    if ctx.obj["json_output"]:
-        click.echo(output.format_json(rows))
-    elif activity:
-        click.echo(output.format_member_list_activity(rows))
-    else:
-        click.echo(output.format_member_list(rows))
-
-
-@member.command("capture")
-@_director_member_options
-@click.option(
-    "--lines",
-    "--tail",
-    "lines",
-    type=int,
-    default=30,
-    show_default=True,
-    help="Lines to capture (alias: --tail).",
-)
-@click.option(
-    "--ansi/--no-ansi",
-    default=False,
-    hidden=True,
-)
-@click.pass_context
-def member_capture(ctx, member_id, lines, ansi):
-    """Capture the last N lines of a member pane's terminal buffer."""
-    _require_fleet_id(ctx)
-    fleet_id = ctx.obj["fleet_id"]
-
-    _ensure_tmux_or_die()
-
-    target, placement = _load_authorized_member(
-        fleet_id,
-        member_id,
-    )
-    member_id = target["agent_id"]
-    pane_id = _require_member_pane(placement, member_id, "capture")
-
-    try:
-        content = MULTIPLEXERS["tmux"].capture_pane(target_pane_id=pane_id, lines=lines)
-    except TmuxError as exc:
-        raise click.ClickException(f"capture failed: {exc}") from exc
-
-    if not ansi:
-        content = output.strip_ansi(content)
-
-    if ctx.obj["json_output"]:
-        click.echo(
-            output.format_json(
-                {
-                    "member_agent_id": member_id,
-                    "pane_id": pane_id,
-                    "lines": lines,
-                    "content": content,
-                },
-            )
-        )
-    else:
-        # color=True preserves ANSI escape sequences on non-TTY sinks (e.g.
-        # CliRunner-captured stdout). Without it, click.echo would re-strip
-        # the escapes the operator just opted into via --ansi.
-        click.echo(content, nl=False, color=True if ansi else None)
-
-
-@member.command("send-input")
-@_director_member_options
-@click.option(
-    "--choice",
-    type=click.IntRange(1, 3),
-    default=None,
-    help="Choice 1/2/3 (xor --freetext).",
-)
-@click.option(
-    "--freetext",
-    type=str,
-    default=None,
-    hidden=True,
-)
-@click.pass_context
-def member_send_input(ctx, member_id, choice, freetext):
-    """Safely forward a restricted keystroke to a member pane."""
-    _require_fleet_id(ctx)
-    fleet_id = ctx.obj["fleet_id"]
-
-    if freetext is not None and freetext.lstrip().startswith("!"):
-        raise click.UsageError(
-            "--freetext may not start with '!' — that triggers the coding agent's "
-            "shell-execution shortcut. Use 'cafleet member exec' for shell dispatch instead."
-        )
-
-    supplied = sum(1 for v in (choice, freetext) if v is not None)
-    if supplied != 1:
-        raise click.UsageError(
-            "--choice and --freetext are mutually exclusive; supply exactly one."
-        )
-
-    if freetext is not None and ("\n" in freetext or "\r" in freetext):
-        raise click.UsageError("free text may not contain newlines.")
-
-    _ensure_tmux_or_die()
-
-    target, placement = _load_authorized_member(
-        fleet_id,
-        member_id,
-    )
-    member_id = target["agent_id"]
-    pane_id = _require_member_pane(placement, member_id, "send")
-
-    try:
-        if choice is not None:
-            MULTIPLEXERS["tmux"].send_choice_key(target_pane_id=pane_id, digit=choice)
-            action, value = "choice", str(choice)
-        else:
-            MULTIPLEXERS["tmux"].send_freetext_and_submit(
-                target_pane_id=pane_id, text=freetext
-            )
-            action, value = "freetext", freetext
-    except TmuxError as exc:
-        raise click.ClickException(f"send failed: {exc}") from exc
-
-    if ctx.obj["json_output"]:
-        click.echo(
-            output.format_json(
-                {
-                    "member_agent_id": member_id,
-                    "pane_id": pane_id,
-                    "action": action,
-                    "value": value,
-                },
-            )
-        )
-    else:
-        label = f"choice {value}" if action == "choice" else "free text"
-        click.echo(f"Sent {label} to member {target['name']} ({pane_id}).")
-
-
-@member.command("exec")
-@_director_member_options
-@click.argument("command")
-@click.pass_context
-def member_exec(ctx, member_id, command):
-    """Dispatch a shell command via the coding agent's `!` shortcut."""
-    _require_fleet_id(ctx)
-    fleet_id = ctx.obj["fleet_id"]
-
-    if "\n" in command or "\r" in command:
-        raise click.UsageError("command may not contain newlines.")
-    if not command.strip():
-        raise click.UsageError("command may not be empty.")
-    command = command.strip()
-
-    _ensure_tmux_or_die()
-
-    target, placement = _load_authorized_member(
-        fleet_id,
-        member_id,
-    )
-    member_id = target["agent_id"]
-    pane_id = _require_member_pane(placement, member_id, "exec")
-
-    try:
-        MULTIPLEXERS["tmux"].send_bash_command(target_pane_id=pane_id, command=command)
-    except TmuxError as exc:
-        raise click.ClickException(f"send failed: {exc}") from exc
-
-    if ctx.obj["json_output"]:
-        click.echo(
-            output.format_json(
-                {
-                    "member_agent_id": member_id,
-                    "pane_id": pane_id,
-                    "command": command,
-                },
-            )
-        )
-    else:
-        click.echo(
-            f"Sent bash command {command!r} to member {target['name']} ({pane_id})."
-        )
-
-
-@member.command("ping")
-@_director_member_options
-@_quiet_flag
-@click.pass_context
-def member_ping(ctx, member_id, quiet):
-    """Inject an inbox-poll keystroke into a member's pane (Director-only)."""
-    _require_fleet_id(ctx)
-    fleet_id = ctx.obj["fleet_id"]
-
-    _ensure_tmux_or_die()
-
-    target, placement = _load_authorized_member(
-        fleet_id,
-        member_id,
-    )
-    member_id = target["agent_id"]
-    pane_id = _require_member_pane(placement, member_id, "ping")
-
-    try:
-        ok = MULTIPLEXERS["tmux"].send_poll_trigger(
-            target_pane_id=pane_id,
-            fleet_id=fleet_id,
-            agent_id=member_id,
-        )
-    except TmuxError as exc:
-        raise click.ClickException(f"send failed: {exc}") from exc
-    if not ok:
-        raise click.ClickException(
-            f"send failed: tmux send-keys did not deliver the poll-trigger "
-            f"keystroke to pane {pane_id}."
-        )
-
-    if ctx.obj["json_output"]:
-        click.echo(
-            output.format_json(
-                {
-                    "member_agent_id": member_id,
-                    "pane_id": pane_id,
-                },
-            )
-        )
-    elif quiet:
-        click.echo(str(member_id))
-    else:
-        click.echo(
-            f"Pinged member {target['name']} ({pane_id}) — poll keystroke dispatched."
-        )
-
-
-if __name__ == "__main__":
-    cli()
